move PaymentFacade/Repository logic from project-base to the framework package
